Add registration-interest/sign-up to routes
We're providing two ways for a user to sign up for registration opening notifications. /registration-interest is for the soft close which will be accessed via the npq registration journey. This route will close, with the only route to registering for notifications being`registration-interest/sign-up` url.

The additional route /registration-interest/sign-up is a link that is going to be provided to the govuk website, as part of the hard-close. This has been added in this PR as both will be released at the same time but will remain open until registrations reopen.