[CN-110] Add “Choose your nursery” pages
There’s no way of ending up redirected to this page yet but if works_in_nursery is set to true in the wizard you’ll be directed to this new interface for selecting your school. It has the exact same data backing it but it is presented as separate.

We use the same data for both rather than showing just schools in the school path and just nurseries in the nursery path. This is because which path you end up going down depends in some cases on your perception of whether you’re a part of a school or a nursery. Where these are the same entity it can be unclear which path someone will pick. To resolve that we let people pick either regardless of the path they take.