Add inside_catchment to funding eligibility calculations
It didn’t make sense to be performing these checks everywhere outside of the funding eligiblity, checking it before you checked actual eligibility. It’s a core part of the calculation and should be alongside the rest of the logic.

This means that separate checks in choose_your_npq#next_step can be removed and anywhere that is doing a funding eligiblity calculation will come to the same result, where as previously you would be rejected in one spot but your application wouldn’t account for your catchment.