Add rake task for updating LeadProviders
This adds in a service called Services::LeadProviders::Updater, this service takes in a newly created LeadProviders::ALL_PROVIDERS constant and uses it to update the list of providers in the database to match this hash.

This can then be used in both the seeds and a rake task for updating the database.