Return applicants to flow when core data changed
This fixes a bug highlighted by sentry:
1. Say you don’t work in a school
2. Say you work in england
3. Say you don’t work in childcare
4. Reach the check your answers page without choosing a school
5. Click change on the do you work in a school question
6. Change your answer to yes
7. Attempt to finish your application

Instead of being asked what school you work at and appropriate other questions it will instead attempt to take you back to the check answers screen after the email entry question. At this stage it hits a 500 error because it doesn’t have a school name to output.

So what I’ve done is made it so that certain key questions that determine flow will direct you to the regular flow instead of the /registration/:step/change flow where changing_answer? is set to true. Instead it behaves as if this is your first time through the flow because you have made a drastic change to your answers.