Only load institution when applicant works there
I uncovered a bug where if you managed to pick an institution but then went back and said you didn’t work in that type of place, it would still use it when determining where to send you after you chose an NPQ.

To resolve this, I made it so that when loading your institution it will check whether you work at that type of institution first before loading it, this means that it will no longer be used in funding eligiblity calculations, as it shouldn’t be.

After doing this I realised that the ChooseYourNPQ page did not need to check if you work in a school or childcare anymore as this will be handled by the eligibility calculations themselves.
Once I removed this I found that it would begin sending non-catchment school/childcare applicants down the "where do you work" path because they had no instiutition which was not correct.
Only catchment applicants should go down that route, to resolve that I introduced a catchment check for deciding whether to send them to the "where do you work" path.

I've also added a spec for the non-school non-childcare path

# Conflicts:
#	spec/features/happy_journeys_spec.rb