Remove repeated code from happy journey JS specs
The specs are repetitive in some areas and not always specifying the
current page makes them a difficult to follow.

This change adds the JourneyAssertionHelper that provides us with methods

* navigate_to_page
* given_i_am_on_page
* now_i_should_be_on_page
* and_the_check_your_answers_page_should_contain

This allows us to remove some repeated lines (axe checking) and gives
the test some hierarchy (via blocks) making it easier to see which
assertions happen on which page.

The actual testing process hasn't changed.