Add work_setting step
Work setting is a new page in the wizard that asks users to identify the
kind of establishment they work in (early years, school, college,
other).

This new page replaces the questions:

* do you work in a school?
* do you work in early years or childcare?

However, we need to maintain consistency with older records so we're
inferring the answer to those questions from the work setting question.
This approach also lets us keep our changes to the flow minimal as we
needn't change the logic throughout.