Add manual validation page for admins
This change adds a page that will allow admins to download CSV files
containing details of people who need to be manually validated and then
upload the outcomes.

It reduces the need for developers to log into production and manually
run commands.

As there are now multiple admin pages the feature specs have been split
out, one for each section and one covering the section as a whole
(ensuring normal users can't see it without logging in as an admin).