Remove unwanted keys rather than set to nil
By overriding the answers we no longer care about with `nil` we're
ending up with three possible states (unset, nil and set) per answer.
This makes testing more complicated than it would be if there were just
two (unset and set).

This change updates `work_setting.rb` so we're deleting the answers
that, due to a change of an earlier option, will no longer be needed
at the end.