Rename the display_…? methods to be descriptive
Previously they described what it was they displayed. This new name (display_no_javascript_fallback_form?) correctly descriptions what type of form is being displayed and most importantly *why*?

Without this change it took me a while to really grok why this section was in place and wondered about removing it. Adding this in clears things up and makes it obvious that this is here for two key reasons:
1) Non-JS browsers
2) Accessibility tools