Move institution identifier validation to method
When using the validates macro with a format, the translation
interpolation expects a `urn` which isn't provided. This could have been
solved by changing the interpolated variable to `value` but I opted to
move the entire validation routine to a single method so it's obvious
what's happening (and in what order)

Now the `validate_institution_identifier` method first checks the value
is in the right format _then_ it checks that the corresponding
establishment exists.