Add tracking pixels to the app
These should help DfE gauge how successful the NPQ advertising is.

'Pixels' were originally requested for Meta, LinkedIn and Twitter. The
provided snippets for Meta and LinkedIn included both JavaScript and
`<img>` pixels; Twitter's only provided JavaScript.

Because the JavaScript then set loads of tracking cookies and did other
shady things, we decided not to include it. It would have required us to
rewrite the privacy policy and extend the cookie prefernces screen so
they could be opted in/out of individually.

Instead we're using the 'fallback' pixels only. As Twitter's snippet
didn't provide a pixel it has been dropped.

The most straightforward solution is to reword the 'Accept cookies'
banner so it covers cookies and tracking. Then we can only show the
tracking pixels to people who have opted in.