Use name instead of full_name for TRA name field
I swapped the specs over to using full_name to match the real parsing but then an issue presented in how response stubbing works which meant I had to use name in the specs, to make the two match we now use name instead of full_name in the real response parsing.

The issue in stubbing is that the default mocked response includes a name field, if you don’t override it you get a phantom value appearing in your stubbed responses. Making sure to use name prevents this.