Only query the ECF API if absolutely necessary
Moving the ECFAPI check for if a user has previously received funding after the local check for if the institution+course is eligible reduces the number of API queries we’re making, speeding up response time for most users.

This was highlighted by several tests now requiring webmock stubs when they previously did not.