Copy feature flags to all users on login
This spec was created to replicate a bug that’s triggered 80 times over about 3 months and leads to a full error page for users who encounter it. This flow seems like such an extreme edge-case that it can’t be the only trigger for this bug.

1. Go through the GAI flow without a verified TRN, return and be removed from the pilot as no TRN is returned.
2. Go through the non-GAI flow until you reach the email input, enter a different email to the one you used on GAI
   for a user that already exists that doesn’t have a TRN on the record. You will be logged in as this user.
3. Enter your TRN on the TRN step, it is required and you cannot proceed without one.
4. Proceed and complete the Choose your NPQ page.
5. Encounter a 500 error.

You encounter a 500 error here because after the Choose your NPQ page a funding eligibility check is performed.
This check requires your TRN.
Since your email was already on a user record when you entered it on the email input page you were logged in as
that user but importantly that user did not have your feature flag ID copied across
(since it’s only copied for new users) and so your feature flags reset to the flags for that user.
This user’s flags more than likely would leave you as a part of the pilot.
Because you were still considered a part of the pilot during the eligibility check the TRN for the user would be
looked up on the user record in the wrong place looking where it would be for GAI users instead of for non-GAI users.
Then it tries to use nil when constructing the URL and fails.
The solution I found was to just copy feature flags during log in for new and existing uses.
This overall is a better approach IMO because it’s not a great idea to have any user suddenly jump from one set of flags to another.
The solution was to copy the feature flags from the session ID to the user even if it isn't a new user being created.