Replace confirmation page with registration
Adds a new step method to the form to check if the application has been
submitted / completed or not. I had to do this step because the current version
relies on a manual click from the user to access the registration page.

Because I moved the following two lines into the check-answers step many tests
has been changed.

```ruby
  wizard.store["submitted"] = true
  wizard.session["clear_tra_login"] = true
```

Also, other tests have changed because of changing the copy of the landing page.
In our tests there is quite a bit of duplicated code on our assertions, I will
explore next if we have a way to dry-it-up.