Create an action that pushes to DockerHub
Previously there was some duplication between the deployment workflows,
each created a separate Docker images and pushed them to Dockerhub.

Before the  move to Azure we want the Docker images to be pushed to
GitHub's container registry (as we'll have fewer sets of credentials to
manage). So, instead of implementing it four times I've moved the Docker
image steps to their own action and called it from each workflow.