Log users with duplicated Get-an-identity-id
Some users have two accounts created in NPQ, one of them has an UID assigned and
the other account it doesn't.

This PR, logs them by checking if the user if there is an existing user with the
same UID that we are about to assign.

These cases must be checked individually and an action should be taken.