Replace map with find_each for reporting
We are debugging issues with Delayed::Job after the migration from PAAS to
Azure. It seems that the jobs related to this report are not finishing in
`delayed_jobs`. We are exploring if this could be related to a memory issue as
we are loading all Application (with some includes) via Active Record.

This PR uses `find_each` which is a lighter approach.