Add snapshot database backup/restore actions
As part of the work to pull NPQ out of the ECF application we want to perform
a number of data integrity checks. In order to do this we ideally need safe
access to production data.

Add a GitHub action and associated workflow to generate a snapshot of the
production database, that we can then use `konduit` to access.

Add a make command to create the `konduit` and update the development database
config with the resulting connection string.