Disable validation of user email
As we moved to GAI, we don't need email validation anymore. With each login, we are receiving most recent user email and we update database accordingly. With validation, we are having a clash when GAI email changed and we can not update GAI user due to having non-GAI account with current GAI email. This change will cause duplicated identities on the ECF side, but it does not happen often and we can easily track those.