Add ECF database connection
We want to be able to pull in ECF models; in order to do this we will need to
connect to the ECF database as well as the NPQ database.

Add ECF database to database.yml and update Makefile to allow easier
connecting to the NPQ and ECF snapshot databases.

Ensure `ApplicationRecord` connects to the primary (NPQ) database.