Deploy migration environment separately
We now have a migration environment in Azure that currently works in the
same way as sandbox - commits/merges to main trigger a deployment which
is pushed to sandbox/migration/production at once.

This isn't what we want for our migration purposes. Instead we have a
new npq-separation feature branch where all migration work/testing will
take place. We want it to be deployed to the migration env whenever
commits/merges happen on it.