Add ability to download orphan reports
When we run the migration we have already queried all the data required to
compile an orphan report.

Save the orphan reports in the Rails cache (not ideal, but the esiest way to
have this work across multiple web instances) and add an endpoint to retrieve
the reports for a given migration.

DRY up mocking the in-memory Rails cache.