Minor performance optimizations
Loading the applications for the users (ECF in particular) takes a _long_ time.
We use this when tentatively matching orphans on school. By touching the
applications in the memoization it means we'll only do this for every user
once, but as there are ~270k users it takes quite a while. If it becomes a
problem we can look at further optimisations, such as a custom subquery.