Optimise user/application queries
Only return the fields we're actually using as part of the reconciliation
process. Currently this only restricts the fields of the main query object;
its proving tricky to limit the fields coming back from included
relationships without resorting to a `joins` instead of `includes`, which
would cause issues elsewhere that we rely on the model structure.