Spike swagger documentation
A spike to generate a Swagger yaml spec and documentation UI using the `rswag`
gem.

I looked at other gems that can do this, but `rswag` appears to be the
defacto-standard and most flexible all-round way of Swaggerizing a Rails
application. nd generating a documentation UI.

`rswag` generates a Swagger yaml file from descriptive specs, which is nice as
it kills two birds with one stone. It should do everything that we would need
and more.

The `rswag-ui` package leverages `swagger-ui` under the hood, which generates
a UI that mostly matches Mark's design and has a decent plugin system for
customizing the various components. It uses `React` and appears to operate as
`Rack` middleware, which is a bit of a pain when it comes to customising the
`Components`. The issue is we can't import `React` with `webpacker` as the
helpers aren't available in middleware. Instead, I've pulled `React` from a
CDN and inlined the plugins to customise the UI. There's probably a better way
of doing this.

Between the plugin system and CSS/custom JS we should be able to change any of
the design we like, though some changes are easier than others. I wasn't able
to hide the `Topbar` component, for example, as it appears to set `Redux`
state for other parts of the page, so I had to hide this with CSS. If we want
to modify a particular `Component` we have to redefine it and pass it in as a
plugin, which is OK for most scenarios but some of the default components are
quite complex and redefining them can be verbose.

Most of the `Redux` state is pulled from the `swagger.yaml` generated in
Rails, but `swagger-ui` also has a way of overriding selectors and reducers so
that we can modify the state if we have to.

Overall I think its a decent solution, as long as we don't go too crazy with
the UI it should be fairly painless to implement.