Add some very basic user to participant sync tasks
Note that here we're keeping the ids the same across users and
participants, this will make it easier to track who's who.

After we run any 'big sync' task we'll need to update the participants
sequence[0] accordingly so it doesn't generate clashing numbers. In
production this would be a one-off.

https://www.postgresql.org/docs/current/sql-altersequence.html