Fix GAI session tests
Here the GAI session spec was logging in a user using the magic link,
which is only used by admins. We're only checking that the links appear
when a user is logged in, so it's more straightforward to stub
`User.find_by`, which means current_user will return a truthy value and
the logout/GAI info will be displayed.

Not perfect but better.