Add login helper for development/staging/review
We currently face some difficulty logging in as an admin in
environments we use for development and review because the email codes
aren't actually sent. Locally we can check the server log where the code
is printed, but on staging/review it's a chore because to see logs we
need PIM.

This change adds a helper which returns nil unless the environment is
staging, development or review, in which case it prints the OTP code.

It's in a helper rather than in the template so we can easily test it,
it would be very bad if this didn't work properly, and the method is
less prone to error than logic directly in a template.