Disable autorestore in dotenv
As of v3 this is enabled by default, but I think one of our gems must
modify `ENV` and it results in an error. Disabling for now, its bad
practice to modify `ENV` directly and we don't do it at the moment.

If we find the guilty gem would could just disable it for a subset of
tests, but its probably not worth the effort to debug it.