Ensure dotenv doesn't run if gem isn't loaded
We don't load the test gems during the docker build process, but we precompile
the assets in the test environment which results in an error as `dotenv` will
autoload if it finds a `.env` file.

Ignore the `.env` files so they don't end up in the docker image. Conditionally
disable autoload.