Connect to the ECF database
We want to be able to connect to the corresponding ECF database in each
environment (apart from review as we don't spin up an ECF database here).

Update database config with ECF connections. Update `ApplicationRecord` to
connect to the NPQ reg database and add a `BaseRecord` for the ECF models. The ECF schema is included so that we can generate a test database.