Add MigrationJob
Add job for kicking off a migration. Split out the preparation logic on the
`Migrator` service so that we can create the `DataMigration` instances in-band
and therefore render the state more easily in the UI (if we did this in the job
itself it may not happen before the page reloads after starting a migration).