Add controller and views for migration
Add the controller actions and views necessary for a user to be able to kick
off, view the progress and outcome of a migration. Only super admins will be
able to run migrations.

Fixes a typo in the `sign_in.html.erb` error message.

Adds a config option for `rspec` to easily run tests against an in-memory cache
and `rack_test` Capybara driver.