Migrate lead providers from ECF to NPQ reg
Update the `Migrator` service to migrate lead providers from ECF to NPQ reg. As
all the lead providers should already exist in NPQ reg, the migration consists
of verifying consistency between the two services. A failure will be recorded
if a lead provider exists in ECF that cannot be found in NPQ reg (using the
`ecf_id` attribute).