Migrate lead providers from ECF to NPQ registration (#1205)
* Fix adding konduit URL to database config

When we moved to multiple databases (to connect to ECF) we should have updated
the indentation on this make command.

* Migrate lead providers from ECF to NPQ reg

Update the `Migrator` service to migrate lead providers from ECF to NPQ reg. As
all the lead providers should already exist in NPQ reg, the migration consists
of verifying consistency between the two services. A failure will be recorded
if a lead provider exists in ECF that cannot be found in NPQ reg (using the
`ecf_id` attribute).