Add swagger-ui to render OpenAPI yaml
Initially I tried adding via the `swagger-ui` package, which would have been
ideal as we would only pull in/package the parts we use, but getting this to
properly transpile with our outdated version of `webpacker` proved more hassle
than its worth and would have polluted our `package.json` with a whole bunch of
extra modules.

Instead, I've pulled in via `swagger-ui-dist` which is a prepackaged version we
can simply include and use.