Move config out of SubNavigationComponent
The config here was previously hardcoded but we now plan on using this
component in both the admin console and the API documentation.

Now the config is hardcoded in a class that can be passed into the nav,
meaning the two things aren't tied together and entirely customisable
and reusable.

https://sandimetz.com/blog/2009/03/21/solid-design-principles