Check OpenAPI schema for changes during CI
We want to make sure the various swagger.yaml files are regenerated when a
change is made. If they aren't it can get out of sync with the code and leads
to issues. We also want to prevent anyone manually modifying the schema files.

Run a checksum on the schema files, then regenerate them from the specs,
recalculate the checksums again and fail the build if they are different.