Ensure parity between serializers and swagger
We want to ensure that when an attribute is added to a serializer the
corresponding documentation exists for it in swagger. At the moment this
process involves manually updating the respective swagger components, however
going forward we may be able to build on Blueprinter to allow us to auto
generate the response models.

Add test coverage so that if an attribute exists in a serializer and not in
the corresponding swagger component (and vice-versa) we hit a failure.

Switch from `field` to `association` in Blueprinter for the attributes so that
we can parse its structure (otherwise we just have a block for the `attributes`
key).