Separate NavigationStructure to its own file
Adding the structure classes to the same files as the containing nav
components was a bit confusing and the two parts (primary and sub),
which are closely tied together were apart.

Now they have been combined into a single NavigationStructure class
which has two public methods, #primary_structure and #sub_structure

This makes it easier to see the navigation tree as a whole. It's a
little verbose but should give us the flexibility we need for even
complex navigation pages.