Make admin navigation components reusable (#1222)
* Remove lambdas from subnav

These were tried when some of the structure was pulled from the database
(so we didn't needlesly run queries when on other pages) but now
everything's static again, just go back to arrays of sections

* Disable some more spurious Rubocop rules

Style/SignalException - no, it's more sensible to use raise when we plan
                        on rescuing and fail when we don't

RSpec/LeadingSubject  - no, this is just annoying, sometimes we use let
                        variables within the subject and in
                        expectations, just let the developer choose an
                        appropriate order

* Move config out of SubNavigationComponent

The config here was previously hardcoded but we now plan on using this
component in both the admin console and the API documentation.

Now the config is hardcoded in a class that can be passed into the nav,
meaning the two things aren't tied together and entirely customisable
and reusable.

https://sandimetz.com/blog/2009/03/21/solid-design-principles

* Move config out of PrimaryNavigationComponent

Similarly to the previous commit, moving the admin structure out of the
primary navigation component makes it reusable. The structures are
defined as subclasses of PrimaryNavigationStructure

* Update admin template to use new nav formats

We're now setting the sub nav's current_section based on the section
from the primary nav, simplifying it considerably and removing some
duplication too.

* Move nav components up from admin directory

Now they're going to be shared they don't need to be within the admin
namespace

* Separate NavigationStructure to its own file

Adding the structure classes to the same files as the containing nav
components was a bit confusing and the two parts (primary and sub),
which are closely tied together were apart.

Now they have been combined into a single NavigationStructure class
which has two public methods, #primary_structure and #sub_structure

This makes it easier to see the navigation tree as a whole. It's a
little verbose but should give us the flexibility we need for even
complex navigation pages.