Add rswag and swagger-ui for API documentation (#1224)
* Add swagger-ui to render OpenAPI yaml

Initially I tried adding via the `swagger-ui` package, which would have been
ideal as we would only pull in/package the parts we use, but getting this to
properly transpile with our outdated version of `webpacker` proved more hassle
than its worth and would have polluted our `package.json` with a whole bunch of
extra modules.

Instead, I've pulled in via `swagger-ui-dist` which is a prepackaged version we
can simply include and use.

* Add rswag to document statements endpoints

Add `rswag` gem, configured to output OpenAPI spec in `yaml` format.

Set up initial swagger docs for the statements endpoints.

Regenerate the API docs with `rake rswag:specs:swaggerize`.

* Check OpenAPI schema for changes during CI

We want to make sure the various swagger.yaml files are regenerated when a
change is made. If they aren't it can get out of sync with the code and leads
to issues. We also want to prevent anyone manually modifying the schema files.

Run a checksum on the schema files, then regenerate them from the specs,
recalculate the checksums again and fail the build if they are different.

* Precompile assets before running tests

If we don't precompile the assets they are compiled as part of the first
feature spec, which can take a long time and cause it to timeout
(swagger-ui-dist is the culprit).

* Ensure parity between serializers and swagger

We want to ensure that when an attribute is added to a serializer the
corresponding documentation exists for it in swagger. At the moment this
process involves manually updating the respective swagger components, however
going forward we may be able to build on Blueprinter to allow us to auto
generate the response models.

Add test coverage so that if an attribute exists in a serializer and not in
the corresponding swagger component (and vice-versa) we hit a failure.

Switch from `field` to `association` in Blueprinter for the attributes so that
we can parse its structure (otherwise we just have a block for the `attributes`
key).

* Add documentation for swagger

* Remove parity checking test

It turns out `rswag` already does this; I tested with nullable/required
attributes and removing/adding on the serializer/schema sides and it always
picked it up.

* Populate swagger URL from environment

* Revert to using sandbox URL for swagger

As the swagger docs are generated locally and not during CI we can't generate
them with the correct URL for the environment they are being deployed to.

* Scope API guidance code block formatting

This was effecting the swagger UI as well as API guidance.