Make email/uid validation conditional
As we have invalid data in the NPQ reg database at the moment that we
can't easily fix, we only want to apply the email validation when validating
logic that is part of the NPQ separation work. Likewise, the `uid` of an
existing user can change via the `UserUpdater` at the moment, so we only want
to scope that validation to particular scenarios.

Change `email` and `uid` validation to apply only with a custom context, for
example: `user.valid?(:npq_separation)`.

Supress Rubocop warning; until we can fix the data and apply the unique
constraint on the database.