Add separation environmnent (#1270)
* Add new Rails env for separation

Add a new Rails environment for the separation work and enable separation
features (apart from migration as there is no corresponding ECF environment to
reference yet). Clean up the `webpacker.yml` file.

* Add terraform for separation environment

Adds terraform configuration for separation environment in the production space.

* Deploy to separation environment on merge to main

We want to deploy to the separation environment when merging to `main`,
alongside the `sandbox` and `migration` environments.

* Update environment documentation

Add documentation for the separation environment. Clean up docs in general.

* Fix migration if provider does not exist

On a fresh installation the `School-Led Network` provider will not exist and the migrations fail.

* Run seeds on separation environment

* Revert webpacker.yml changes

This isn't right, but its how its currently setup and for some reason asset
precompilation doesn't work for these environments. One for another day.