Support database refreshing in monitoring checks
We have a `migration` environment where we can trigger a database refresh from
the `production` database via a GitHub action at any time. As the monitoring
checks to see if the database is populated this causes the instance to become
unhealthy whilst refreshing the database.

We want to allow the database to be empty for the `migration` environment to
allow for this.

General monitoring house keeping and test coverage improvements.