Add missing fields to Application from ECF
Adds missing fields to the `Application` model from ECF.

We need to make the `ecf_id` on the `users` table unique in order to have it as
a foreign key for the `elgiible_for_funding_udpated_by` relationship.

It also looks like the `ecf_id` on the existing NPQ reg models is a `text`
field whereas it has a `uuid` type in ECF. We should probably update all these
in a separate PR to be consistent.