Add Redis cache for migration environment (#1284)
We are using the cache in the migration environment to store failure reports
for the migration runs. At the moment these persist to an in-memory cache, so
we can't access it from the application worker (as the cache is written from a
job worker).

Add Redis for a shared cache that we can reliably query after writing from any
instance.