Add ApplicationSerializer
Ports the `ApplicationSerializer` and respective versions from ECF. The only
field we can't currently return correctly is `schedule_identifier` as we
haven't migrated schdules yet.

Instead of inheriting between serializers as we do in ECF we're leveraging
Blueprinter views to be able to return version-specific fields in the
serializers.

Move existing `StatementSerializer` out of the version-specific module.