Add missing fields to Application from ECF (#1289)
* Add missing fields to Application from ECF

Adds missing fields to the `Application` model from ECF.

We need to make the `ecf_id` on the `users` table unique in order to have it as
a foreign key for the `elgiible_for_funding_udpated_by` relationship.

It also looks like the `ecf_id` on the existing NPQ reg models is a `text`
field whereas it has a `uuid` type in ECF. We should probably update all these
in a separate PR to be consistent.

* Remove `eligible_for_funding_updated_x` fields from Application

We don't appear to surface these anywhere apart from the admin portal at the
moment and we can't easily add them due to needing `ecf_id` to be a unique
index on users and it currently contains `null` values.