Add schools/courses to admin dashboard (#1314)
* Add schools to the admin dashboard

Add a page to list schools and show details of a school in the admin dashboard.

For some reason the Faker town/county translations are missing and it
complains, so I've hardecoded these values in the factory for now.

* Add courses to the admin dashboard

Add a page to list courses and show details of a course in the admin dashboard.

* Switch ordering of columns on schools admin page

On feedback from Mark I'm switching these around.