Combine `Statements::Find` and `Statements::Query` (#1311)
* Remove Statements::Find class

* Make statement method take ecf_id or id

ecf_id will be predominantly used by the API but internally we'll use
id. It makes sense to refer to them explicitly by name to avoid
confusion.

* Allow Statements::Query to query by state

* Replace references to Statements Find with Query

* Remove enforced output_fee scope

The argument defaults to true to keep the existing (common) behaviour
but can be overridden if necessary.

* Add support for querying with arrays

This is more likely when searching internally rather than via the API so
it's a good idea to support both styles.