Add application endpoint (#1305)
* Add rebranded_alternative_courses to Course

This is used when working out if an Application has been previously funded.

Improve test coverage in `Course` model.

* Add funding status logic to Application

When we serialize an application we need to specify a
`ineligible_for_funding_reason`, which we determine based on if the application
has been previously funded and is currently eligible for funding.

Add `previously_funded?` and `ineligible_for_funding_reason` to `Application`
(lifted from ECF).

* Add ApplicationSerializer

Ports the `ApplicationSerializer` and respective versions from ECF. The only
field we can't currently return correctly is `schedule_identifier` as we
haven't migrated schdules yet.

Instead of inheriting between serializers as we do in ECF we're leveraging
Blueprinter views to be able to return version-specific fields in the
serializers.

Move existing `StatementSerializer` out of the version-specific module.

* Add Applications::Query service

Add a query service to find applications, ordered by `created_at: :asc`, which
mirrors the ECF query service.

Only the filtering used by the application endpoint has been implemented; the
rest can be added as and when we add the endpoints that use them.

Remove `Find` service and replace usage with `Query`.

* Fix declaration spec file paths

These were incorrectly hitting the application endpoints.

* Default application factory to have ecf_id

We have this as a trait currently but it makes more sense for applications to
get it by default as the vast majority will be in this state.

* Add application show endpoint

Add endpoint to retrieve an application. Rename specs to remove `_controller`
suffix to match statements. Extract specs into a shared context to DRY up
between versions/endpoints.

* Fix teacher_catchment serialization

We should be returning `true` only if they are in a UK catchment area.

* Add type to ApplicationSerializer

Add type field and reorder the fields to match ECF.

* Support `id` and `ecf_id` when query applications

We will likely need to support querying by both `id` and `ecf_id` in this
service.