Introduce sentinel default values
The previous implementation of the statements query had the slight flaw
that the default values for the keyword query arguments were `nil`, and
`nil` is meaningful in ActiveRecord, which converts it to `null`.

Instead if we use the `:ignore` symbol to mark arguments as 'omitted',
it frees up `nil` for `null` queries.

This change also shifts the building of the scope to the initializer
and allows us to just keep the scope attr instead of one for each
argument.

Almost inadvertantly, this format combined with the `:ignore` pattern
brings back chaining, which might be useful as we progress.